Post-hoc analysis of the sample-scores
I have added scripts to run linear or logistic regression models to test for association the sample-gene-wide-scores, script Sample_score.R

The script run_glmnet_SampleScore.R run a lasso regression with all genes in the input. it will spit out the regression coefficients and two pseudo R-square measures, that of Efron and of the McKelvey and Zavoina, which can be used as a proxy for the variance explained.

Signed-off-by: inti <intipedroso@gmail.com>