Change in gene stats
The original formula of the combined z-score uses the 
square of the (z^2)/V. This removes the sign of z and generates
probles when z is very negative, i.e. very insignificant. 
to void this problems wi have change the formula and new also returns
z/sqrt(V). the output file reports this stat (a z-score from a N(0,1)) and it p-value

Signed-off-by: inti <intipedroso@gmail.com>